+++ CHANGELOG.md
@@ -8,7 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     * Hide datepicker window on scroll.
 
 * `object-list-view` component:
-    * Toggle selectAll buttons title.
+    * Toggle title of checkAll buttons.
 
 ### Fixed
 * The state of the move buttons in the `flexberry-groupedit` now depends on the selected rows.